+++ src/main/java/ru/gruzoff/entity/OrderDetails.java
@@ -19,7 +19,7 @@ public class OrderDetails extends BaseEntity {
     protected Adress adressTo;
 
     protected Date dateTime;
-    protected int timeOnOrder;
+    protected float timeOnOrder;
 
     @Column(columnDefinition = "TEXT")
     protected String comment;
